Cleanup refactoring
Factor out get_{remove,add}_path_methods
Factor out secret "write_env" fn
Run rustfmt on everything
